api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.1.1
python-dotenv==1.0.0
httpx>=0.23.0,<0.25.0
pydantic>=2.0.0 